# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.1"
  title: portal para los amantes del futbol
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /helloworld:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /gamesystem:
    x-swagger-router-controller: gamesystem.controller
    get:
      description: get the game system list
      operationId: getEstadiosfutbols
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetGameSystemListResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse" 

    post:
      description: add a new game system to the list
      operationId: addEstadiosfutbols
      parameters:
        - in: body
          name: gamesystem
          description: The Game System to create.
          schema:
            $ref: "#/definitions/GameSystem"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetGameSystemResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"


  /gamesystem/{id}:
    x-swagger-router-controller: gamesystem.controller
    get:
      description: get a game system
      operationId: getEstadiosfutbolsbyId
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetGameSystemResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a game system
      operationId: updateEstadiosfutbols
      # define the parameters
      parameters:
        - name: id
          description: game system id
          type: number
          in: path
          required: true
        - in: body
          name: gamesystem
          description: The Game System to update.
          schema:
            $ref: "#/definitions/GameSystem"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetGameSystemResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
    
    delete:
      description: delete a game system
      operationId: deleteEstadiosfutbols

      # define the parameters
      parameters:
        - name: id
          description: game system id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"    
   
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  GameSystem:
    type: object
    properties:
      
      stadium:
        type: string
        description: Description of the Game System
      city:
        type: string
        description: Name of the Game System
      capacity:
        type: integer
        description: Name of the Game System
      createdAt: 
        type: string
        description: Name of the Game System
      updatedAt: 
        type: string
        description: Name of the Game System
        
        
    
  GetGameSystemResponse:
    type: object
    properties:
    
      stadium:
        type: string
        description: Description of the Game System
      city:
        type: string
        description: Name of the Game System
      capacity:
        type: integer
        description: Name of the Game System
   
  GetGameSystemListResponse:
    required:
      - gamesystems
    properties:
      gamesystems:
        type: array
        items:
          $ref: "#/definitions/GetGameSystemResponse"
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
     
      
 
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string